Update omp mem stress test
