+++ func/omp/CMakeLists.txt
@@ -29,11 +29,11 @@ omp_func(custom_reduce custom_reduce.cpp)
 omp_func(for_static_schedule for_static_schedule.cpp)
 omp_func(header_api_support header_api_support.cpp)
 omp_func(hellomp hellomp.cpp)
+omp_func(mem_stress mem_stress.cpp)
 omp_func(mt_pi mt_pi.cpp)
 omp_func(nested_levels_test nested_levels_test.cpp)
 omp_func(nested_parallel nested_parallel.cpp)
 omp_func(omp_checks omp_checks.cpp)
-omp_func(omp_malloc omp_malloc.cpp)
 omp_func(reduction_average reduction_average.cpp)
 omp_func(reduction_integral reduction_integral.cpp)
 omp_func(setting_num_threads setting_num_threads.cpp)
